Added 'monitor' command, which controls stereo/mono monitoring
The monitor command controls the 'direct monitoring' mode of the
input pairs on the interface.

The inputs are paired as 1+2 and 3+4, and each pair can be
configured as stereo or mono. If a pair is configured as mono
then both of those physical inputs will be heard equally in the
left and right outputs (either headphones, or line 1-4), whereas
if a pair is configured as stereo, then the odd-numbered inputs
will only be heard through the odd numbered outputs, likewise the
even inputs and even outputs.